Catch exceptions from interceptor on Event.Join()
Leaking exceptions to scala actor library will cause the error to be
silently eaten by the actor library, and cannot be reported to the user
of the ccf library. Also, server might be called with synchronous actor
calls (server !? Event.Something()) and we always want to return either
Event.Ok() or Event.Error().

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>