Ignore reply messages from transport in server
Because of the way scala actors work, message sent using reply() in the
transport in reply to our asynchronously sent Event.Msg() will end up in
the server's normal actor mailbox. This might cause a never-ending loop
because of both parties always replying to Event.Error() with
Event.Error(). To prevent this situation, we need to just ignore the
messages for now. In the future, a better solution should be crafted,
but this fix will prevent these problems for now.

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>