Use double-bang for sending asynchronous messages to transport
By using the double-bang (!!) operator in sending asynchronous messages
to the transport, an anonymous reply channel will be created which will
silently "eat" the reply message, instead of it ending up into the
Server's inbox. This way we don't have to handle the replies from the
transport, because we are not really interested in them anyway.

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>