Initialize transport when the server starts
In order to improve initialization procedure and communication protocol
between the transport and the server, we want to give the transport as
constructor parameter to the server, which then initializes the
transport with a reference to itself. This way the references to neither
server nor transport needs to be delivered in the actor messages, and
both parties can take the initiative in the communication and the
message-protocol between them can be kept cleaner.

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>