Introduce initial version of the test application
This simple collaborative text editing application will be developed
along with the ccf to serve as an example project and demonstration of
the usage of the ccf library. Currently it's not in a great shape and
has no tests whatsoever, but it will be developed and it's better to
have it in the same repository to make synchronizing the development
easier.

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>