app: Transfer messages between client and server instead of ops
Now the flow between server and client contains messages, and the client
is responsible for managing it's state using
JupiterOperationSynchronizer. This prevents the client from being
implemented directly using JavaScript or other similar methods, because
the transformation function need to be available and implemented in the
client too. However, this approach better suits for ous example
application, as this is the way CCF will most probably be used (at least
in it's current form).

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>