app: Use ccf server instead of direct usage of Jupiter synchronization
Use the actor-based ccf server class to handle the synchronization in
the example application also. For now the code is quite clumsy and is in
the need of obvious refactorings, but it's better than the previous one.

Seems that the instantiation of even a simplest ccf server requires
currently maybe a little bit too much boilerplate code. We might need so
simplify the framework initialization code somehow.

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>