Fix double join after timouted join returns unrecoverable error
If a join operation timeouts it might be executed on the server, but
another join to the same channel will be executed directly after it. The
client doesn't receive the result/state from the first join and the
second join results in an error. The only way around this problem is to
return the same backlog on re-join to same channel otherwise the client
will crash on timeouted joins.

Signed-off-by: Nuutti Rintala <nuutti.rintala@reaktor.fi>