transport: Replace Headers with Map[String, String]
Currently, cumbersome constructs such as request.headers.headers are
needed to properly format data to wire format. This problem stems from
the fact that Headers is a wrapper around TreeMap, and does not
implement the Map trait.

Extending collection classes in Scala is difficult and error prone, so
instead of extending existing collection, such as HashMap, Headers is
replaced with bare Map[String, String].

The strict element order preservation is not necessary, so simple
immutable Map is enough.

Signed-off-by: Harri Salokorpi <harri.salokorpi@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>