transport: Update interface of HttpClient
Instead of allowing the user to pass a function that parses the response
data, simply return a String that can be processed by the user. The main
motivation for such modification is the impact on test cases that cover
code invoking HttpClient#post. These tests must correctly match
invocation including the passed function and this makes testing even
harder.

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Acked-by: Harri Salokorpi <harri.salokorpi@reaktor.fi>