session: Add PartRequest
Since the part operation is meant for parting the channel, the channel
id is the actual request payload and is thus contained in the content
instead of header.

Operations such as sending a request to channel should store the channel
id in the header, since in that case it should be considered as a part
of the request metadata.

Signed-off-by: Harri Salokorpi <harri.salokorpi@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>