session: Add Message
All session layer messages must extend Message. The purpose of the
Message is to make all operations support retrying and queueing.

Message#send unifies transmission of messages among different message
types, which allows re-sends and queuing of messages. This requires
that message specific sending logic is placed in #send.

Re-transmission of a message is at the user's discretion based on an
error condition or detected inconsistency in sequence ids.

Signed-off-by: Harri Salokorpi <harri.salokorpi@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>