session: Add Join message
Join message checks the session channel list and discards the join
message if channel is already joined.

The reason for placing the channel check to Join instead of Session is
to keep the handling of different messages on Session as uniform as
possible. By doing the check in Message, the Session does not have to
handle the re-transmission and sequence verification of Join message
separately from, say, Part message or application-level message.

Signed-off-by: Harri Salokorpi <harri.salokorpi@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>