build: Upgrade sbt from version 0.5.5 to 0.7.3
Quite useful PathFinder#flatMap is missing in version 0.5.5, which is
needed when creating a distribution Zip file. Of course, one can copy
all files to 'dist' directory and then zip the content of this
directory, however, the resulting code is far from concise.

Also, dropped overrides of 'testFrameworks' as obviously useless:

  $ ./sbt
  > project ccf
  > console-project
  scala> testFrameworks.map(.implClassName)
  res1: Seq[String] = List(...)

Signed-off-by: Karim Osman <karim.osman@reaktor.fi>
Acked-by: Harri Salokorpi <harri.salokorpi@reaktor.fi>