perftest: Add simple Python web server for performance testing
In order to profile bottlenecks in CCF's HTTP communication, a simple
Python webserver comes very handy. One can reason that the existing HTTP
server in CCF does not meet the performance requirements, if the
perftest client is clearly faster against the Python web server.

The web server will respond with same response despite of the request
path or request type, and can be started up with:

 > python webserver.py

Signed-off-by: Harri Salokorpi <harri.salokorpi@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>