perftest: Use Operation messages in performance testing
Join and Part messages update the set of Session. Perftest currently
generates a lot of join requests to random channels, which causes the
channel set to grow beyond expected size. The set handling in this case
will most likely start to slow down the request processing, skewing the
results.

Use of Operation messages for testing removes the potential result skew.

Signed-off-by: Harri Salokorpi <harri.salokorpi@reaktor.fi>
Signed-off-by: Karim Osman <karim.osman@reaktor.fi>