session: Add tests to verify session state
Verify that list of session channels is updated correctly when joining
or parting a channel through SessionActor mechanisms.  In order for us
to inspect the internal state of the session actor, SessionActor now
returns the current state when a shutdown is requested, so we can
inspect the end state.

Streamline also the creation of sessionactors in the specification by
creating the instance in the context creation instead individually in
each test case.

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>