session: Add OperationContextMessage
OperationContextMessage requires a template argument for the operation
type, because OperationContext is always templated with the given
operation type. Hence OperationContextMessage can not be an object (it't
not possible to create templated objects in scala). To unify the
creation of messages, all message subclasses were changed to case
classes instead of objects, and the construction of messages is now done
with #create().

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>