session: Split construction of session requests to factory class
SessionRequest.scala was growing a bit too long, so construction is now
split into a factory class instead of inlining it to session request.
Some of the tests are still kept in SessionRequestSpec.scala because of
dependencies to local data structures there. However, we should move the
testing of creation to factory spec at some point while implementing the
actual creation.

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>