server, perftest: Change ServerEngine to perform also transport decode & encode
ServerEngine now performs decoding of request from String and encoding of
response to String for servers.

Decoding and Encoding is done by trait transport.Codec that is passed to
ServerEngine as constructor parameter.

ServerEngine does not yet process requests in any way, it simply replies with
success response to all requests. No error handling is implemented either.
All errors are propagated as exceptions to the client.

Signed-off-by: Sami Rosendahl <sami.rosendahl@reaktor.fi>