transport: Use HttpTransportHeaderContributor in HttpConnection
When given a contributor, call it in #send and make sure given headers
are added. Allows implementing e.g. authentication by adding the
authentication token using HttpTransportHeaderContributor in the client
side and then checking it in the server using
TransportRequestInterceptor.

Current test case is a bit ugly, containing dependencies to the Json
encoder/decoder in order to assert the sent request correctly. The
internal structure of HttpConnection would need some cleanups in order
to make it more testable.

Signed-off-by: Aki Saarinen <aki.saarinen@reaktor.fi>